+++ pyphare/pyphare/pharein/simulation.py
@@ -1,7 +1,6 @@
 import os
 
 from ..core import phare_utilities
-#from .globals import objects
 from . import globals
 
 def compute_dimension(cells):
