+++ pyphare/pyphare/pharesee/__init__.py
@@ -1,1 +1,1 @@
-from . import box
+
